+++ media/shaders/glsl/lambert_v.glsl
@@ -10,7 +10,7 @@ attribute vec2  Texcoord0;
 
 #if defined( SKINNED )
 attribute vec4  BlendWeights;
-attribute ivec4 BlendIndices;
+attribute vec4  BlendIndices;
 #endif
 
 #if defined( USE_COLOR )
