+++ media/shaders/glsl/phong_v.glsl
@@ -13,7 +13,7 @@ attribute vec2  Texcoord0;
 
 #if defined( SKINNED )
 attribute vec4  BlendWeights;
-attribute ivec4 BlendIndices;
+attribute vec4  BlendIndices;
 #endif
 
 #if defined( USE_COLOR )
